LAPIC: Fix issue where the LAPIC would not be enabled in Application Processors.
The issue was coming from the way the kernel enables the LAPIC: Through
the IA32_APIC_BASE register (by setting bit 11). However it seems that
this only work for the BSP and not the APs.

This fix uses the 8th bit in the Spurious Interrupt Vector register of
the LAPIC to enable it, for the BSP and APs. This is much more reliable.