AP wake up procedure: Fix issue where APs would not release the lock on the boot stack due to an off-by-one.
The issue arose when adding the target field in the data frame. At some
point in the wake up routine, the target function pointer was saved into
the stack to be preserved while switch to paging. After restoring the
value, the ESP was not changed and the instruction supposing to unlock
the boot stack was writing to the wrong location on the stack.

Added a test in smp to boot APs with a single stack.