Disable BSP bit in the IA32_APIC_BASE MSR of the BSP upon boot.
According to Intel's manual, a processor having the BSP bit set in its
IA32_APIC_BASE will not react to INIT Inter-Processor-Interrupt the same
way an AP does. More specifically it will not enter a wait-for-SIPI
state.
Because of this, the original BSP cannot be "rebooted" by an AP using an
INIT-SIPI-SIPI sequence.
Even though in theory this is not a big limitation (since there is no
point in rebooting the BSP), in practice this makes init_aps() testing
impossible from APs.

Hence disable the BSP bit in IA32_APIC_BASE MSR so that the BSP can be
rebooted by APs when they are running init_aps() tests.