Handle case where the BSP has a non 0 APIC ID.
The Intel specification does not mention that the BSP chosen at power-up
or RESET must have an APIC ID of 0. Therefore, it would be possible to
end up in a situation where the BSP has a non-0 APIC ID. Make the
existing AP wake up routine compatible with this scenario.

This change also makes possible to test the init_aps() functions from
APs.