Add new mechanism of global and local callbacks for interrupt vectors.
Global callbacks are global to all cpus, while cpu local are, well,
local to the cpu.  When an interrupt is raised, the generic interrupt
handler look for a local callback for the given vector. If such a
callback is present, the handler calls the callback and irets. If no
local callback is registered for that vector, the handler looks for a
global callback and, if found, calls it.  If there is no local nor
global callback registered for the vector a PANIC is raised.