Fix bug in LAPIC implementation where the Interrupt Command Register was incorrectly mapped.
The higher 32 bits of the ICR where not mapped to the correct physical
address meaning that the destination field was not written properly.
This was causing issues when trying to send an IPI to a specific cpu
(normal delivery mode with no destination shorthand) where the
destination field was not written and thus was always zero leading to
the BSP receiving all IPIs.