Kmalloc: Improve dynamic memory de-allocation by de-allocating a group once it becomes empty.
After de-allocating a particular address, the memory allocator checks if
the group that contained that address became empty after this
de-allocation. If so, then the group is removed from the group list and
is itself de-allocated.
De-allocation of a group is important as it frees up the underlying
physical frame(s) used by the group.