Inter-Processor-Messaging: Now free allocated struct ipm_message_t.
A new field in struct ipm_message_t has been added: receiver_dealloc.
This boolean indicates if the receiver of a message is responsible to
de-allocated the message upon processing it.

The reason the receiver of a given message is responsible to de-alloc it
is because the sender has no idea when a particular message has been
processed or not (technically it could look at the remote's message
queue). Instead of the sender waiting around for the remote to process a
message, the remote can be told to free the message instead. This is
more efficient.

This field can also be set to false, although there are no use cases for
such a thing now, this could be useful if a message is ever statically
allocated.