Kmalloc: Fix fragmentation issue.
This fragmentation would occur in two scenarios:
Scenario 1: A free() occurs while a cpu creates a new group:
    1. CPU 0 calls do_kmalloc. No group can contain the allocation.
    2. CPU 0 releases the KMALLOC_LOCK and calls create_group.
    3. CPU 1 calls free() on a piece of memory bigger than CPU 0's
       allocation.
    4. CPU 0 re-acquire the lock, insert the new group in the list and
       allocate in it.
This is fixed by making sure, after re-acquire the KMALLOC_LOCK that an
allocation does not fit in any of the existing group. If it does then
there is no need to add the new group and we can re-try the allocation.

Scenario 2:
    1. CPU 0 calls do_kmalloc. No group can contain the allocation.
    2. CPU 0 releases the KMALLOC_LOCK and calls create_group.
    3. CPU 1 calls do_kmalloc. No group can contain the allocation.
    3. CPU 1 releases the KMALLOC_LOCK and calls create_group.
    4. CPU 0 re-acquire the lock, insert the new group in the list and
       allocate in it.
    5. CPU 1 re-acquire the lock, insert the new group in the list and
       allocate in it.
In this scenario we end up with two group with low utilization. The fix
is the same as Scenario 1: After re-acquiring KMALLOC_LOCK, re-try the
allocation before inserting the new group. In this case CPU 1 should
detect that the allocation is successful (in the group added by CPU 0)
and therefore should not add the new group.