Paging: Refactor paging functions so that paging operation can be performed on arbitrary address spaces.
The initial implementation of paging function would only mutate the
address space currently used on the cpu. Now function takes an
additional struct addr_space * parameter indicating the address space
that should be modified.

Function operating on arbitrary address space have an "_in" suffix. Each
of these function also have a macro with the same name minus the suffix
which goal is to provide an easy way to call the functions with the
current address space as target. This also avoid changing all the
existing calls to paging functions.

Note: Testing has not been done yet on calling these functions on
different address spaces. Moreover physical frame leaks are expected
when calling these functions on address spaces that are different than
the current one. This will be taken care of in a subsequent submit.