Interrupts: Improve register saving upon interrupts.
The _interrupt_handler_common routine handling interrupts in a generic
manner (at low level) is now responsible to save _all_ the registers of
the interrupted context. This information is passed through the struct
interrupt_frame to the generic_interrupt_hander() function.

This is the first step towards process scheduling. For now registers
restoration upon an iret is only limited to EIP, EFLAGS and the 8
general purpose registers. This ok as interrupt contexts are limited to
kernel code for now.