Scheduler: Add per cpu runqueues.
The scheduler now picks the first process it find in the current cpu's
runqueue or default to the idle_proc.
Once a proc has been selected to be run, it is removed from the runqeue.
Upon selecting a new process to run the old curr_proc is enqueued back
on the cpu's runqueue.