Scheduling: Rework on the scheduling logic.
This commit introduces a new struct sched which provide a generic
interface of a scheduler (i.e. function pointers). A new scheduler
implementation only needs to define those function pointers.

The other new mechanism introduced by this commit is the scheduling core
logic (defined in sched_core.c). This core logic is a scheduling policy
independent logic, this is the layer above the struct sched.

This design is heavily inspired from the linux kernel.

This commit also introduce the TS (Trivial Scheduler) struct sched as an
example.