USTAR: Add implementation of the USTAR filesystem.
Implementation of the USTAR filesystem. This filesystem is no more than a read
only TAR archive. It is used for the init ram-disk.

The format of a TAR file (and hence of this filesystem) is straightforward:
     HEADER | DATA * | HEADER | DATA * | ....
Everyfile is preceded by a header (defined by struct ustar_header) immediately
followed by its data. Everything (header and data) are organized in 512 bytes
blocks. The length of a file is rounded up to the next multiple of 512. Files
might not have data block (if their size is 0), in this case their header is
followed immediately by the header of the next file.  Directories are
equivalent to empty files.