Disk: Add support for memdisks.
memdisks are virtual disks that reside in RAM. Just like physical disk
one can read and write sectors to memdisks. Memdisk must have a size
that is a multiple of 512 bytes.

memdisks can be created using the create_memdisk() function which takes
an address and a length of a memory area to be used as a virtual disk.
They can be deleted with delete_memdisk().