VFS: Add first implementation of the Virtual File System (VFS).
This is a _very_ simple implementation of VFS which is in no way near
what Linux's VFS is capable of. The goal of this VFS is to be able to
mount disks at certain path and make opening, closing, deleting, reading
from and writing to files easier. The current implementation of VFS has
undefined behaviour if the same file is opened twice and written and
read concurrently (or written twice).