Interrupt: Fix issue where nested interrupts would corrupt the saved register values of the current process.
The bug was due to the fact that the interrupt handler would assume the
interrupted context was the current process' context. This assumption
would fail with nested interrupts which can occur if a cpu sends a TLB
message within an interrupt handler and waits for the acks with
interrupts enabled.

The fix is to add a counter per cpu that counts the level of nested
interrupts. If an interrupt is nested then the interrupt handler will
skip the scheduler and return to the interrupted context directly.