Scheduling: sched_start() now resets the interrupt_nest_level before running the very first process.
This is meant to avoid issues if this function is called using a remote
call. If this is the case and the interrupt_nest_level is not reset then
the very first interrupt received by the cpu will be considered nested
and the scheduler will be skipped.

This also fix a bug in the scheduler test because the exact scenario
above was happening.