VFS: Rework on the assumption on the mapping between open files and struct file's. There is now at most a single struct file per absolute path, this means that if a file A is opened multiple times (in multiple or same process), a single struct file for A is instantiated and shared amongst the processes using it.
This is implemented using a linked list in VFS containing all the opened
files (system wide) called the Opened Files Linked List (OFLL). Upon
calling vfs_open(), the function looks up the file in the OFLL and, if
found, returns the struct file associated with the pathname.
struct file now contains a reference count "open_ref_count" meant to
indicate how many processes/fds are sharing this instance. When 0 the
struct file can be removed from the OFLL and freed.