IPM: Fix memory leaks with remote function calls for functions that do not return.
This issue was particularly big in testing where most remote functions
calls do not return and instead wait for a RESET.
The fix is to make sure that the ipm_message and remote_call_data
structures are freed before running a function. This is done in the case
the sender is not waiting for completion.