Syscall/do_exit(): Fix bug in do_exit() syscall where the syscall implementation would call sched_run_next_proc() which does not return. This meant that do_exit() would not return and hence the interrupt_nest_level would not be decreased (this is done at the end of the generic interrupt handler). Because of this, all interrupts after a do_exit() would be considered nested.
The fix is to simply not call sched_run_next_proc() and return from
do_exit() to eventually the generic interrupt handler. Since do_exit()
is called by a process though a syscall we know that the nesting level
is 1 and hence the generic interrupt handler will trigger a round of
scheduling. Since the process is marked as dead the scheduler will pick
another one.