Paging: Fix BUG in create_temp_mapping() where the previous temporary mapping for the cpu was not disabled which lead to a PANIC in map_page_int() when mapping the temp address.
The bug was due to the fact that the address of the cpu's temp mapping
entry in the temp mapping page table was wrongly computed. Therefore a
given cpu would write garbage to other PTE instead of unsetting the
present bit in its own PTE.
This behavior was only experienced by cpu for which the ID is > 0.
Regression test added.