Proc: Avoid copying process registers to struct proc for every interrupt.
The struct proc's registers_save field has been changed to a pointer
pointing to the location on the proc's kernel stack where the saved
register values have been pushed by the interrupt handler. This field
was also renamed saved_registers.

When creating a process, the initial values of saved registers is stored
at the bottom of the kernel stack.