Segmentation/Boot: Rework the Boot GDT so that the kernel can execute into higher half _logical_ addresses during initialization. This allows the kernel to access global variables (which addresses are linked into higher half) even before enabling paging.
This is possible by setting the base of the code and data segments of
the Boot GDT to -KERNEL_PHY_OFFSET, that is -0xC0000000 = 0x40000000.
This way, when accessing a global variable which has been linked to
virtual address 0xCABCDEF0:
    logical address = 0xCABCDEF0
    linear address = 0xCABCDEF0 + 0x40000000 = 0xABCDEF0 = physical
        address where the variable has been loaded.

By doing so, the initialization code can access any global variables as
soon as the Boot GDT is loaded (in init_segmenation, which is the first
function called in the entry point of the kernel). Additionally, logging
can be enabled at that time as well and is hence available when
initializing paging.

Once paging has been initialized but before the ID mapping is removed
(higher half mapping and ID mapping are coexisting in init_paging()),
the base of the code and data segments of the Boot GDT are reverted to
0x0, effectively making them flat segments. At this point, the higher
half mapping is provided by paging and the ID mapping can be removed.

Not only this change enables accessing global state almost everywhere,
but it also allows us to get rid of all the shenanigans where subsystems
that could work both before and after enabling paging would need to
check if paging has been enabled and if not call to_phys() on the
address of some global variable (ex: the frame allocator with the
FRAME_BITMAP). Another advantage of this change is that init_paging()
does not need to call all the fixup functions for esp, ebp and the
return address of the function, since all of those are already higher
half.