ACPI: Fix BUG where cpus that are not online capable would be counted towards the total number of cpus on the system. This would mislead the kernel into thinking there are more cpus than the system actually has.
This issue occurs in baremetal situations where the ACPI tables sometime
report more cpus than the system possesses. Those extra CPUs would have
their flags set to 0 in the MADT, meaning that they are not currently
enabled and are not online capable.

The fix is to count only the cpus that are enabled at the time of ACPI
parsing or online capable.