SMP: Rework/Refactor SMP core wake up procedure.
The ap_entry_point routine has been rewritten from scratch to be more
efficient and without the AP needing a stack. This means that there are
no more shenanigans where small real-mode stacks would be allocated and
shared between APs (requiring locks and complexity). The APs now wake-up
in real time and won't have a stack until paging is enabled. This allows
the BSP to allocate the higher-half stacks in advance (before waking up
the APs) and pass them through the data frame.