Refactor interrupt.c, add support for Task-Gate descriptors in the IDT, add Task-Gate for Double Fault exceptions in order to catch kernel stack overflows.
Refactor interrupt.c:
Get rid of init_desc() function, it has been replaced by the IDT_ENTRY()
macro which builds a union interrupt_descriptor_t given a segment
selector, and offset and a DPL.

Task-Gate descriptors:
New macro IDT_TASK_GATE_ENTRY() similar to IDT_ENTRY() for task-gate
descriptors.

Double-Fault interrupt task:
The kernel now uses a Task-Gate for Double Fault exception. This is done
by adding a new TSS entry in the GDT and using a task-gate entry for
vector 0x8 in the IDT pointing to the new TSS.
The reason behind using a task for #DF interrupts is to be able to catch
kernel stack overflows when a cpu tries to write on its stack canary
page. Since at that point the cpu cannot push interrupt info onto the
stack, the cpu will double fault. The double fault handler _MUST_ be a
task so that the cpu can switch to a clean stack to handle the
interrupt.
For now the interrupt callback handling double faults simply prints the
EIP of the double fault in the logs and PANICs. This is ok as #DF
interrupts require a reset anyway.