Interrupts: Fix issue with Double Fault interrupt handler where the busy bit of the TSS segment descriptor was not reset when a cpu would execute the handler.
This meant that if another cpu was trying to access the handler, the cpu
would see the busy bit be 1 and would triple fault.
The "fix" is to reset the busy bit as soon as the interrupt handler
executes. This fix is not perfect since it leaves a window between the
time a cpu jumps to the interrupt handler/task and the time it resets
the busy bit, during which another cpu could try to execute the
interrupt handler/task.
In practice this race condition is expected to be extremely rare since
it would require two cpu overflowing their kernel stack at the same
time.

This commit also adds a test for detecting kernel stack overflows.