Processes: Refactor of struct proc, context switch and process creation.
Struct proc:
The saved_registers field of the struct proc has been removed and
kernel_registers has been renamed to registers. The rationale is that
this registers field contains the value of the registers at the time of
the last context switch that paused the execution of the process.

Context switch:
The function performing the context switch now supports both switching to
a user context (ring 3) and switching to another kernel context. Upon
resuming execution, all the registers are restored, except for CS, DS,
SS and GS for kernel procs.
Added test for function saving the registers of a process.

Process creation:
Since the function performing the context switch now supports switching
to a user context directly, it is not necessary to create a stack frame
to fake a call to initial_ret_to_proc() to do the first IRET to user
space when a user process starts execution for the first time. This
removes a lot of complexity during process creation.