Processes: Fix BUG where the ESP0 of the TSS of the current cpu would not be changed prior to executing a user process.
This bug meant that user processes would user the kernel stack of the
cpu instead of their own kernel stack when handling interrupts. This
would lead to corruption when multiple user processes would run on the
same cpu.

This bug was due to the recent refactor of the proc.c function, this
refactor mistakenly removed the call to change_tss_esp() before
branching execution to a user process.

The fix is to add this call prior to do_context_switch(). This commit
adds a regression test.

This commits also adds a new function to the segmentation.h:
get_tss_esp0(uint8_t const cpu) which allows one to read the value of
the ESP0 field of the TSS of a cpu. This is required for the regression
test since cpus cannot read their TSS outside of segmentation.c.