Processes/Testing: Add test for context switches.
In this test we are running N processes on the same cpu and performing
2N context switches. This test makes sure that registers are not changed
between context switches.

We mix user processes and kernel processes. There is no scheduler
running, hence the test is explicitely using switch_to_proc() to perform
the context switches.
All N programs are executing the same code (at different priv levels):
 1. INT 0x80. This is the very first instruction, no register is
    changed.
 2. Save registers into the struct proc_test_data present on its stack
    (ESP is assumed to point to that struct).
 3. Write 1 to the flag field of the struct.
 4. INT 0x80 again.
The test uses a special handler for 0x80 which will perform a context
switch every time it is called. Hence the expected interleaving of the
processes look as follows:
 * Proc 1 starts executing and execute the INT 0x80 instruction.
 * The handler for 0x80 performs a context switch.
 * Proc 2 starts executing and execute the INT 0x80 instruction.
 * The handler for 0x80 performs a context switch.
     ...
 * Proc N starts executing and execute the INT 0x80 instruction.
 * The handler for 0x80 performs a context switch.
 * Proc 1 resumes execution after the INT 0x80 and save its registers
   onto its stack. Finally it executes a INT 0x80 again.
 * The handler for 0x80 performs a context switch.
 * Proc 2 resumes execution after the INT 0x80 and save its registers
   onto its stack. Finally it executes a INT 0x80 again.
 * The handler for 0x80 performs a context switch.
     ...
 * Proc N resumes execution after the INT 0x80 and save its registers
   onto its stack. Finally it executes a INT 0x80 again.
 * The handler terminates the execution of all processes.
After that the cpu running the test will compare the values of the
register reported by the processes (through their stack) with the start
values. They should match except for EIP (which should point to after
the INT 0x80 instruction).