Scheduling: Fix BUG where a process migrating from one cpu to another could lead to a corruption of its kernel stack.
The corruption was due to the fact that, when a cpu (A) called
put_prev_proc on its current process, it would still use its kernel
stack. At this point, another cpu (B) could see the process into the
scheduler and pick it as its next process (through pick_next_proc). When
resuming the execution, B would use the process' kernel stack to build
the IRET frame and return to the execution. However, during this time A
is still using the kernel stack of the process! Hence the stack
corruption.

To fix this issue, the put_prev_proc has been moved to the routine
performing the context switch, more specifically in resume_proc_exec. A
core in resume_proc_exec will now switch to the kernel stack of the
`next` process no matter if it is a kernel or user process and then
call put_prev_proc on the previous process.

This commits adds a regression test for this bug.