Rework context switching.
This commit change context switching so that it happens by simply
switching to the kernel stack of the next process to run. The previous
implementation would save the entire context of the current process
(e.g. all registers), restore the context of the next, and execute an
IRET (after building an entire interrupt frame) to wherever the `next`
process stopped last.

While the previous implementation had the added bonus of supporting both
switches to user and kernel processes directly (e.g. the context switch
would jump directly to user space instead of first kernel and then
user), it had a few shortcomings:
        1) Slow. Saving and restoring registers can be done efficiently
           (PUSHA and POPA) but doing an IRET for every context switch
           can become expensive.
        2) Convoluted. Code saving and restoring the state was more
           complex that other solutions.
        3) Because of the IRET, the kernel would jump directly to the
           next context without the opportunity to do some eventual
           bookeeping.

Reason #3 what was led to this change: when switching to the next
process the kernel would not be able to restore preemptions (that need
to be disabled while performing the context switch) unless by jumping
through some hoops (the IRET could jump to a common routine and then
jump to the process' context).

Hence revert to a simpler implementation where:
        1) All context switches under the switch_to_proc() call (more
           precisely under do_context_switch).
        2) A context switch is a simple change of kernel stack.
        3) Since do_context_switch is a function we only need to save
           callee-saved registers.

The only downside to this implementation is that the first context
switch of a process gets more complicated. Upon creating a process the
kernel needs to mock the stack frame of do_context_switch() so that the
process execute the correct code when it executes for the first time and
hits the RET at the end of do_context_switch. In the case of user
processes, the do_context_switch needs to return to a function that will
prepare an interrupt stack frame and IRET from it to user space.

Note: Some tests had to be disabled since they don't work with kernel
processes anymore. They will be changed and enabled back in a subsequent
commit.